requirements.txt
fastapi>=0.110.0,<1
uvicorn[standard]>=0.23.0,<1

pandas>=1.5.0,<3
numpy>=1.23.0,<3
scikit-learn>=1.2.0,<2
requests>=2.31.0,<3

sentence-transformers>=2.6.0,<4
python-dotenv>=1.0.0,<2
